+++ src/stdlib.rs
@@ -599,6 +599,12 @@ pub fn make_std_mod(config: &Configuration) -> Result<Program, Errors> {
         None,
         Some(include_str!("./docs/std_iostate_unsafe_create.md").to_string()),
     ));
+    errors.eat_err(fix_module.add_global_value(
+        FullName::from_strs(&[STD_NAME, FFI_NAME, DESTRUCTOR_NAME], "_make"),
+        destructor_make(),
+        None,
+        Some(include_str!("./docs/std_ffi_destructor_make.md").to_string()),
+    ));
 
     errors.to_result()?;
     Ok(fix_module)